[finished] binaryheap implementation

+++ Code/binaryheap.py
@@ -3,7 +3,7 @@
 
 class BinaryMinHeap(object):
     """BinaryMinHeap: a partially ordered collection with efficient methods to
-    insert new items in partial order and to access and remove its minimum item.
+    insert new items in partial order and to access and remove its minimum item
     Items are stored in a dynamic array that implicitly represents a complete
     binary tree with root node at index 0 and last leaf node at index n-1."""
 
@@ -21,8 +21,7 @@ class BinaryMinHeap(object):
 
     def is_empty(self):
         """Return True if this heap is empty, or False otherwise."""
-        # TODO: Check if empty based on how many items are in the list
-        # ...
+        return self.size == 0
 
     def size(self):
         """Return the number of items in this heap."""
@@ -94,10 +93,17 @@ class BinaryMinHeap(object):
         # Get the parent's index and value
         parent_index = self._parent_index(index)
         parent_item = self.items[parent_index]
-        # TODO: Swap this item with parent item if values are out of order
-        # ...
-        # TODO: Recursively bubble up again if necessary
-        # ...
+        
+        # Swap this item with parent item if values are out of order
+        # since this is mean heap the parent has to be smaller or equal
+        # to its children
+        if parent_item > item:
+            self.items[index], self.items[parent_index] = (
+                self.items[parent_index], self.items[index])
+            # recursively call if heap is still out of order
+            # if index != 0:
+            index = self._parent_index(index)
+            self._bubble_up(index)
 
     def _bubble_down(self, index):
         """Ensure the heap ordering property is true below the given index,
@@ -105,23 +111,33 @@ class BinaryMinHeap(object):
         Best case running time: O(1) if item is smaller than both child items.
         Worst case running time: O(log n) if items on path down to a leaf are
         out of order. Maximum path length in complete binary tree is log n."""
+        # print(f"items: {self.items}")
         if not (0 <= index <= self._last_index()):
             raise IndexError('Invalid index: {}'.format(index))
         # Get the index of the item's left and right children
         left_index = self._left_child_index(index)
         right_index = self._right_child_index(index)
-        if left_index > self._last_index():
+        # check for if left and right index are in bound
+        if left_index > self._last_index() or right_index >= self.size():
             return  # This index is a leaf node (does not have any children)
         # Get the item's value
         item = self.items[index]
-        # TODO: Determine which child item to compare this node's item to
+        # Determine which child item to compare this node's item to
         child_index = 0
-        # ...
-        # TODO: Swap this item with a child item if values are out of order
+        left_child = self.items[left_index]
+        # print(f"leftchild: {left_child}")
+        right_child = self.items[right_index]
+        # print(f"rightchild: {right_child}")
+      
+        child_index = left_index if left_child < right_child else right_index
+
+        # Swap this item with a child item if values are out of order
         child_item = self.items[child_index]
-        # ...
-        # TODO: Recursively bubble down again if necessary
-        # ...
+        
+        if item > child_item:
+            self.items[index], self.items[child_index] = self.items[child_index], self.items[index]
+            # Recursively bubble down again if necessary
+            self._bubble_down(child_index)
 
     def _last_index(self):
         """Return the last valid index in the underlying array of items."""
@@ -159,12 +175,14 @@ def test_binary_min_heap():
         correct = heap_min == real_min
         print('get_min: {}, correct: {}'.format(heap_min, correct))
 
-    print('\nDeleting items:')
-    for item in sorted(items):
-        heap_min = heap.delete_min()
-        print('delete_min: {}'.format(heap_min))
-        print('heap: {}'.format(heap))
-        print('size: {}'.format(heap.size()))
+    delete = True
+    if delete:
+        print('\nDeleting items:')
+        for item in sorted(items):
+            heap_min = heap.delete_min()
+            print('delete_min: {}'.format(heap_min))
+            print('heap: {}'.format(heap))
+            print('size: {}'.format(heap.size()))
 
 
 if __name__ == '__main__':
